### 2020-09-04 #2640 @whedon
[PRE REVIEW]: BrainGlobe Atlas API: a common interface for neuroanatomical atlases
**Submitting author:** @adamltyson (<a href="http://orcid.org/0000-0003-3225-1130">Adam Luke Tyson</a>)
**Repository:** <a href="https://github.com/brainglobe/bg-atlasapi" target ="_blank">https://github.com/brainglobe/bg-atlasapi</a>
**Version:** 0.2.0
