### 2020-09-04 #2640 @adamltyson
Is the PDF failing to compile because I gave a link to the repository branch? I couldn't work out how to specify the branch (with the paper on) when submitting.
