### 2020-09-04 #2640 @kthyng
Hi @adamltyson! Thank you for your submission. Is `https://github.com/brainglobe/bg-atlasapi/tree/joss` the url of the repository you want reviewed as your submission?
