### 2020-09-04 #2640 @kthyng
Hi @oliviaguest, would you be up for editing this paper, once the repo/paper situation has been worked out?
