### 2020-09-04 #2640 @adamltyson
Hi @kthyng, the repository is `https://github.com/brainglobe/bg-atlasapi`. `https://github.com/brainglobe/bg-atlasapi/tree/joss` is the `joss` branch that is otherwise identical to the default, but with the paper. I wasn't sure how to submit the URL to reflect this.
