### 2020-09-04 #2640 @kthyng
Hi @openjournals/dev! Do I have the power to change the URL in a submission in the database, rather than just editing it in the comment above? I think what needs to happen is for the repo url to change to "https://github.com/brainglobe/bg-atlasapi" with the JOSS paper then being accessible as a branch from there.
