### 2020-09-07 #2640 @xuanxu
> Do I have the power to change the URL in a submission in the database, rather than just editing it in the comment above? I think what needs to happen is for the repo url to change to "https://github.com/brainglobe/bg-atlasapi" with the JOSS paper then being accessible as a branch from there.

@kthyng (sorry for the delayed response) There is no database, Whedon gets the information from the issue body, so changing the URL there (in the **repository** entry) should make it work.
