### 2020-09-08 #2640 @kthyng
> @kthyng (sorry for the delayed response) There is no database, Whedon gets the information from the issue body, so changing the URL there (in the **repository** entry) should make it work.

Ok thanks @xuanxu, that makes sense. @oliviaguest thanks for updating accordingly.
