### 2020-09-13 #2668 @oliviaguest
@openjournals/dev why is the check references command not working? Is it related to the fact it's on a custom branch?
