+++ paper/paper.bib
@@ -34,6 +34,7 @@ ISSN={1662-5196}
               Joshua T and Gerfen, Charles R and Hantman, Adam W and Korff,
               Wyatt and Sternson, Scott M and Spruston, Nelson and Svoboda,
               Karel and Chandrashekar, Jayaram",
+  doi      = {10.1016/j.cell.2019.07.042},
   journal  = "Cell",
   volume   =  179,
   number   =  1,
