### 2020-09-13 #2668 @FedeClaudi
Hi @vitay  I've added the DOI:  https://github.com/brainglobe/bg-atlasapi/commit/dbe6a38ab3a8d692cfd16f74c88e086347fee1d2, thank you for pointing that out.
