### 2020-09-14 #2668 @vitay
I guess I am already done with the review:
* `bg-atlas-api` is a software with a very clear rationale, providing a simple centralized access to various neuroatanomical atlases. It makes mostly sense within the brainglobe suite, but its API is generic enough so that other tools can build on it.
* The software respects all standards and good practice: hosting on github, BSD license, unittests, CI on travis, online documentation, community guidelines, etc. I have not read the complete code, but the Python code is readable and respects conventions.
* Installation works out of the box with pip (tested with python 3.7 and 3.8), with only a handful of dependencies.
* The promised functionality is there. I have had troubles installing the other elements of the suite (e.g. brainrender-gui) but I could interact with the library directly, using the provided notebook as a basis.
*  The companion paper correctly describes the scope of the software.

In short, I recommend acceptance.
