+++ tests/test_core_atlas.py
@@ -13,7 +13,7 @@ def test_initialization(atlas):
     assert atlas.metadata == {
         "name": "example_mouse",
         "citation": "Wang et al 2020, https://doi.org/10.1016/j.cell.2020.04.007",
-        "atlas_link": "http://www.brain-map.org.com",
+        "atlas_link": "http://www.brain-map.org",
         "species": "Mus musculus",
         "symmetric": True,
         "resolution": [100.0, 100.0, 100.0],
