### 2020-09-29 #2668 @typically
Okay, done reviewing. In summary, I love the tool, especially its simplicity and extensibility. I could definitely see how this could be integrated into any number of projects, including my own.

- Nice simple tool
- Excellent documentation
- Well organized Github page
- Easy to install (pip 20.1.1)
- Paper is readable and makes a good case for need

There are good basic examples of how to use the tool, but as minor suggestions:

1. It might be useful to demonstrate with a practical example how to deal with coordinates (Cartesion, image indices). There is a reference to another package, but it isn't super clear to me how I would go from a downloaded atlas to some standard coordinate system.
2. Hierarchical information appears to be provided (in the form of "structure_id_path"), and discussed in the paper. It would be useful to provide an example of how to use this information to obtain parcellations at different levels of hierarchy. It is not clear to me how this information can be represented, as the "annotation" image contains only a single integer per voxel.

I also recommend acceptance.
