### 2020-09-29 brainglobe/bg-atlasapi#76 @FedeClaudi
Add examples/tutorials for structures hierarchy
As suggested [here](https://github.com/openjournals/joss-reviews/issues/2668#issuecomment-700623351), we could add:

- [x] tutorials to show how to print/visualise the structures hierarchy for a given atlas (i.e. how to  use [the .hierarchy](https://github.com/brainglobe/bg-atlasapi/blob/e8c5d5164b6c1a2a56e5359de53cf5d3f4c53bf2/bg_atlasapi/core.py#L89))
- [x] tutorials to show how to get the children/parent for a given brain region (e.g. [how to use these methods](https://github.com/brainglobe/bg-atlasapi/blob/2d8c1210f3b3cdc5a54bee173877c6f5d3accd63/bg_atlasapi/core.py#L250))

cc @adamltyson  @vigji
