### 2020-09-29 #2668 @adamltyson
Hi @typically, thanks for your review!

For your first suggestion, I'm not 100% clear what you meant, do you mean something like:
1) How the voxel spacing in  the images corresponds to the "real world"? i.e. showing that you can query the different resolution atlases in microns, and get the same results? And also show that the generated meshes for each brain region are in micron units (not voxels).

or

2) How to get your own data into the standard coordinate system? e.g. using [brainreg](https://github.com/brainglobe/brainreg) to register sample image data to the atlas template image, and then analyse your data in the common coordinate space?
