### 2020-09-29 #2668 @typically
Hi @adamltyson, I'm thinking of human templates, but more generally stereotaxic spaces. How would one map an image to Cartesian coordinates such as MNI-152, Talairach space, or rat stereotaxic coordinates, relative to bregma? You would need an origin and a basis vector, I assume. Just a suggestion though!
