### 2020-09-30 #2668 @adamltyson
Thanks for clarifying @typically. We don't have any method for converting between coordinate systems, but this would certainly be useful. Currently we only have one coordinate system for each species (e.g. each of the mouse atlases uses the same coordinates). If there's interest (and data available) for atlases with additional coordinate spaces, we could also provide methods to transform coordinate between coordinate spaces.

So, if it's alright with you, suggestion 1 will go onto our todo list, but it will depend on which atlases we add. We will document suggestion 2, as these methods exist in the software, but are not fully documented yet.

@oliviaguest, we'll add these small changes to the tutorials, then I think we'll be ready to publish.
