+++ requirements.txt
@@ -8,3 +8,4 @@ click
 rich
 tqdm>=4.46.1
 bg-space>=0.5.0
+pyinspect
