### 2020-09-30 93269c4
Improving hierarchy usage and tutorial
