+++ setup.cfg
@@ -12,7 +12,7 @@ search = __version__ = "{current_version}"
 replace = __version__ = "{new_version}"
 
 [flake8]
-ignore = E203,W503,E501,E731,C901,W291,W293,E741
+ignore = E203,W503,E501,E731,C901,W291,W293,E741,F401
 max-line-length = 79
 max-complexity = 18
 exclude = __init__.py
