### 2020-09-30 2a5e6ac
Hierarchy improvements (#77)
* Improving hierarchy usage and tutorial

* blacked

* Fixed F401 flake8 error arising b/c of pyinspect

* Test fix

* Test fix
