### 2020-09-30 brainglobe/bg-atlasapi#76 @vigji
This should have been addressed in #77 now
