### 2020-09-30 #2668 @vigji
@typically now both the [tutorial notebook](https://github.com/brainglobe/bg-atlasapi/blob/master/tutorials/Atlas%20API%20usage.ipynb) and the `Usage` section of the docs [here](https://docs.brainglobe.info) discuss more diffusely the hierarchy. As for the mapping from one atlas to the other, pure translations of the origin and changes in resolution and axes order can be done using [https://github.com/brainglobe/bg-space](url), but we currently don't provide general support from affine/nonaffine transformations between spaces.
