+++ paper/paper.md
@@ -55,4 +55,7 @@ Currently six atlases across three species (larval zebrafish, mouse and human) a
 `BG-atlasAPI`'s flexible infrastructure already proved crucial in the development and extension of two software tools for use in neuroscience: `brainreg` [@Tyson:2020] for 3D registration of image data between sample and atlas coordinate space and `brainrender` [@Claudi:2020] for 3D visualisation of both user-generated data and atlas data.
 We hope that other developers will use the API, and develop tools that can be used across neuroscience and other research fields, increasing their reach, and preventing duplication of effort.
 
+# Acknowledgments
+We would like to thank Nouwar Mokayes for the assistance in packaging the Max Planck Zebrafish Brain Atlas within `BG-atlasAPI`. This work was funded by The Gatsby Charitable Foundation, The Wellcome Trust (214333/Z/18/Z) (T.W.M.), by an Investigator Award from The Wellcome Trust (096436/B/11/Z)(T.W.M.) and by the Deutsche Forschungsgemeinschaft (DFG,German Research Foundation) under Germany’s Excellence Strategy within the framework of the Munich Cluster for Systems Neurology (EXC 2145 SyNergy – ID 390857198).
+
 # References
