+++ paper/paper.bib
@@ -163,9 +163,7 @@ Video Abstract
 }
 
 @software{Tyson:2020,
-  author       = {Adam Tyson and
-                  Charly Rousseau and
-                  Troy Margrie},
+  author       = {Tyson, Adam L. and Rousseau, Charly V. and Margrie, Troy W.},
   title        = {{brainreg: automated 3D brain registration with
                    support for multiple species and atlases}},
   month        = aug,
