### 2020-10-02 #2668 @adamltyson
We've made some small edits to the paper (affiliations, acknowledgments, reference formatting), and we're now good to publish. @oliviaguest what comes next?
