### 2020-10-03 3ef730c
Linting fix and removed pyinspect for now
