+++ bg_atlasapi/bg_atlas.py
@@ -1,7 +1,6 @@
 from pathlib import Path
 import tarfile
 import requests
-import pyinspect  # for better tracebacks
 
 from rich import print as rprint
 
