### 2020-10-03 #2668 @oliviaguest
@adamltyson perfect — firstly, you need to deposit the code, so create an archive (on Zenodo, figshare, or other) and post the archive DOI here.
