+++ bg_atlasapi/__init__.py
@@ -1,5 +1,5 @@
 __author__ = """brainglobe"""
-__version__ = "0.2.0"
+__version__ = "1.0.0"
 
 
 from bg_atlasapi.bg_atlas import BrainGlobeAtlas
