### 2020-10-04 897aac4
Revert "Bump version: 0.2.0 → 1.0.0"
This reverts commit ee7cc78d550892a1a943360c9d7715c5d01657a2.
