+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 1.0.0
+current_version = 0.2.0
 commit = True
 tag = True
 
