### 2020-10-04 e216371
update travis (#79)
