+++ .travis.yml
@@ -11,11 +11,11 @@ stages:
 jobs:
   include:
     - stage: test
-      name: "Ubuntu Xenial python 3.6"
+      name: "Ubuntu python 3.6"
       language: python
       os: linux
       env: MINICONDA_URL=http://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
-      dist: xenial
+      dist: bionic
       python: 3.6
       before_install: source travis/install_unix.sh
       script: bash travis/test.sh
@@ -24,39 +24,39 @@ jobs:
         - coveralls
 
     - stage: test
-      name: "Ubuntu Xenial python 3.7"
+      name: "Ubuntu python 3.7"
       language: python
       os: linux
       env: MINICONDA_URL=http://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
-      dist: xenial
+      dist: bionic
       python: 3.7
       before_install: source travis/install_unix.sh
       script: bash travis/test.sh
 
     - stage: test
-      name: "Ubuntu Xenial python 3.8"
+      name: "Ubuntu python 3.8"
       language: python
       os: linux
       env: MINICONDA_URL=http://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
-      dist: xenial
+      dist: bionic
       python: 3.8
       before_install: source travis/install_unix.sh
       script: bash travis/test.sh
 
     - stage: test
-      name: "Windows Python 3.7"
+      name: "Windows Python 3.8"
       language: shell
-      env: TRAVIS_PYTHON_VERSION=3.7
+      env: TRAVIS_PYTHON_VERSION=3.8
       os: windows
       before_install: source travis/install_windows.sh
       script: bash travis/test.sh
 
     - stage: test
-      name: "macOS Python 3.7"
+      name: "macOS Python 3.8"
       os: osx
       language: shell
       env:
-      - TRAVIS_PYTHON_VERSION=3.7
+      - TRAVIS_PYTHON_VERSION=3.8
       - MINICONDA_URL=https://repo.anaconda.com/miniconda/Miniconda3-latest-MacOSX-x86_64.sh
       before_install: source travis/install_unix.sh
       script: bash travis/test.sh
@@ -66,15 +66,15 @@ jobs:
       language: python
       os: linux
       env: MINICONDA_URL=http://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
-      dist: xenial
-      python: 3.7
+      dist: bionic
+      python: 3.8
       before_install: source travis/install_unix.sh
       script: bash travis/lint.sh
 
     - stage: deploy_test
-      name: "Windows Python 3.6"
+      name: "Windows Python 3.7"
       language: shell
-      env: TRAVIS_PYTHON_VERSION=3.6
+      env: TRAVIS_PYTHON_VERSION=3.7
       os: windows
       before_install: source travis/install_windows.sh
       script: bash travis/test.sh
@@ -88,7 +88,7 @@ jobs:
       script: bash travis/test.sh
 
     - stage: deploy_test
-      name: "macOS Python 3.6"
+      name: "macOS Python 3.7"
       os: osx
       language: shell
       env:
@@ -104,6 +104,16 @@ jobs:
       env:
       - TRAVIS_PYTHON_VERSION=3.8
       - MINICONDA_URL=https://repo.anaconda.com/miniconda/Miniconda3-latest-MacOSX-x86_64.sh
+      before_install: source travis/install_unix.sh
+      script: bash travis/test.sh
+
+    - stage: deploy_test
+      name: "Ubuntu Xenial python 3.8"
+      language: python
+      os: linux
[…]
