### 2020-10-04 4dcfa0d
move lint tests to separate stage
