+++ .travis.yml
@@ -1,4 +1,5 @@
 stages:
+  - lint
   - test
   - name: deploy_test
     if:
@@ -10,6 +11,16 @@ stages:
 
 jobs:
   include:
+    - stage: lint
+      name: "Lint"
+      language: python
+      os: linux
+      env: MINICONDA_URL=http://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
+      dist: bionic
+      python: 3.8
+      before_install: source travis/install_unix.sh
+      script: bash travis/lint.sh
+
     - stage: test
       name: "Ubuntu python 3.6"
       language: python
@@ -60,16 +71,6 @@ jobs:
       - MINICONDA_URL=https://repo.anaconda.com/miniconda/Miniconda3-latest-MacOSX-x86_64.sh
       before_install: source travis/install_unix.sh
       script: bash travis/test.sh
-
-    - stage: test
-      name: "Lint"
-      language: python
-      os: linux
-      env: MINICONDA_URL=http://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
-      dist: bionic
-      python: 3.8
-      before_install: source travis/install_unix.sh
-      script: bash travis/lint.sh
 
     - stage: deploy_test
       name: "Windows Python 3.7"
