+++ .travis.yml
@@ -81,29 +81,11 @@ jobs:
       script: bash travis/test.sh
 
     - stage: deploy_test
-      name: "Windows Python 3.8"
-      language: shell
-      env: TRAVIS_PYTHON_VERSION=3.8
-      os: windows
-      before_install: source travis/install_windows.sh
-      script: bash travis/test.sh
-
-    - stage: deploy_test
       name: "macOS Python 3.7"
       os: osx
       language: shell
       env:
       - TRAVIS_PYTHON_VERSION=3.6
-      - MINICONDA_URL=https://repo.anaconda.com/miniconda/Miniconda3-latest-MacOSX-x86_64.sh
-      before_install: source travis/install_unix.sh
-      script: bash travis/test.sh
-
-    - stage: deploy_test
-      name: "macOS Python 3.8"
-      os: osx
-      language: shell
-      env:
-      - TRAVIS_PYTHON_VERSION=3.8
       - MINICONDA_URL=https://repo.anaconda.com/miniconda/Miniconda3-latest-MacOSX-x86_64.sh
       before_install: source travis/install_unix.sh
       script: bash travis/test.sh
