### 2020-10-04 ed82876
update deployment test ubuntu os
