### 2020-10-04 3a4a274
Update testing (#80)
* update travis

* update testing matrix

* move lint tests to separate stage

* remove double testing

* update deployment test ubuntu os
