### 2020-10-05 #2668 @oliviaguest
Great! @adamltyson I think the title of the repo on zenodo has to be the same as the title of the JOSS paper — change that when you get a moment. 😊
