+++ paper/paper.md
@@ -43,7 +43,7 @@ The BrainGlobe atlas API (BG-Atlas API) overcomes this problem by providing a co
 # Statement of need 
 To facilitate the study of neural function, a long-standing approach has been to identify neuroanatomically defined brain regions: structures with defined function, connectivity and anatomical location. The study of these brain regions led to the development of a number of brain atlases for various species. Typically these atlases are made up of a reference image of a brain, voxel-wise annotations (e.g. a mapping from each voxel to a brain structure) and additional metadata such as region hierarchy (region A is a subdivision of region B). These atlases are used throughout neuroscience, for teaching, visualisation of data, and registration of imaging data to a common coordinate space.
 
-Many excellent and open access atlases exist, such as the Allen Mouse Brain Common Coordinate Framework [@Wang:2020] and the Max Planck Larval Zebrafish Atlas [@Kunst:2019], from which the neuroscience community benefits enormously. These atlases provide a valuable resource for individual scientist and enabled important open-science projects such as Janelia Campus' Mouse Light project [@Winnubst:2019].  Furthermore, for several atlases stand-alone software is available that can be used to explore the atlas' data and requires no coding experience, thus making the atlases accessible to a broader audience. 
+Many excellent and open access atlases exist, such as the Allen Mouse Brain Common Coordinate Framework [@Wang:2020] and the Max Planck Larval Zebrafish Atlas [@Kunst:2019], from which the neuroscience community benefits enormously. These atlases provide a valuable resource for individual scientists and enabled important open-science projects such as Janelia Campus' Mouse Light project [@Winnubst:2019].  Furthermore, for several atlases stand-alone software is available that can be used to explore the atlas' data and requires no coding experience, thus making the atlases accessible to a broader audience. 
 However, to be used in the context of new software (e.g. new visualization tools, or brain registration pipelines)  it is necessary that atlases expose their data through an API. 
 Several commonly used atlases come with APIs, but learning how to use each of them is a time-consuming endavour and can require considerable coding experience. For this reason, often developers produce software that works only with a specific atlas. 
 A single and well documented API that worked across atlases would thus lower the cost of developing new software, which can also be made available for a larger number of scientists. An effort in this direction has been made in the R ecosystem  with the `natverse` package [@Shakeel-Bates:2020], but, to our knowledge, no such option exists in Python, which is emerging as the programming language of choice in neuroscience [@Muller:2015].
