### 2020-10-05 #2668 @adamltyson
Awesome, thanks @oliviaguest, @typically, @vitay!

@arfon - We've been very (very) impressed with the JOSS process, thanks!
