### 2020-10-05 #2668 @FedeClaudi
Yes, thank you very much to everyone involved.
And I can confirm, the submission process as JOSS was very impressive!
