### 2020-10-05 #2668 @arfon
@adamltyson @FedeClaudi - your paper is now accepted into JOSS :zap::rocket::boom:
