### 2020-10-05 #2668 @adamltyson
Hi @arfon, I just noticed the link to the repo on the JOSS page links to the branch with the paper on (that won't be updated), rather than the link at the top of this issue ([https://github.com/brainglobe/bg-atlasapi/](https://github.com/brainglobe/bg-atlasapi/)). Is it possible for this to be changed?
